requirements.txt
python-telegram-bot
selenium
beautifulsoup4
pyinstaller
python-dotenv
requests
psutil
